Create Check if a given graph is tree or not(I think this is a graph question so go thru it when u r visiting graphs)
Write a function that returns true if a given undirected graph is a tree and false otherwise. For example, the following graph is a tree.

Approach 1:
An undirected graph is a tree if it has the following properties. 
There is no cycle. 
The graph is connected.
For an undirected graph, we can either use BFS or DFS to detect the above two properties.