Create Detect a negative cycle in a Graph | (Bellman Ford)
Input : 4 4
        0 1 1
        1 2 -1
        2 3 -1
        3 0 -1

Output : Yes
The graph contains a negative cycle.
cycle

The idea is to use Bellman-Ford Algorithm. 
Below is an algorithm to find if there is a negative weight cycle reachable from the given source.
Initialize distances from the source to all vertices as infinite and distance to the source itself as 0. Create an array dist[] of size |V| with all values as infinite except dist[src] where src is the source vertex.
This step calculates the shortest distances. Do the following |V|-1 times where |V| is the number of vertices in the given graph. 
Do the following for each edge u-v.
If dist[v] > dist[u] + weight of edge uv, then update dist[v]. 
dist[v] = dist[u] + weight of edge uv.
This step reports if there is a negative weight cycle in the graph. Do the following for each edge u-v  
If dist[v] > dist[u] + weight of edge uv, then the “Graph has a negative weight cycle”